backend/requirements.txt
# Core Python Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Deep Learning and Machine Learning
tensorflow>=2.6.0
keras>=2.6.0
joblib>=1.1.0

# Web and API Interactions
requests>=2.26.0
websocket-client>=1.2.0

# Data Visualization (Optional but Recommended)
matplotlib>=3.4.0
seaborn>=0.11.0

# Database (Optional)
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # For PostgreSQL
pymongo>=3.12.0  # For MongoDB

# Crypto Exchange API Libraries
python-binance>=1.0.0
python-kraken-sdk>=0.1.0
python-kucoin>=2.1.0

# Environment and Configuration
python-dotenv>=0.19.0
pyyaml>=5.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
unittest2>=1.1.0

# Development and Debugging
ipython>=7.29.0
jupyter>=1.0.0
pylint>=2.8.0

# Performance and Optimization
numba>=0.54.0
dask>=2021.10.0

# Optional but Recommended for Advanced Features
ta-lib>=0.4.20  # Technical Analysis Library
ccxt>=1.50.0   # Unified cryptocurrency exchange library

# Logging
loguru>=0.5.0

# Security
cryptography>=3.4.0

# Deployment (Optional)
gunicorn>=20.1.0  # WSGI HTTP Server